build: move deploying to start fork scripts, add dai/defo getting to deployment
