+++ deploy/localhost/000_init.ts
@@ -1,7 +1,7 @@
 import { DeployFunction } from "hardhat-deploy/types";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 
-import { chainName, deployInfo } from "../utils/helpers";
+import { chainName, deployInfo } from "../../utils/helpers";
 
 const func: DeployFunction = async (hre: HardhatRuntimeEnvironment) => {
   const { getNamedAccounts, getChainId } = hre;
