+++ deploy/localhost/999_post_checks.ts
@@ -1,13 +1,10 @@
 import { DeployFunction } from "hardhat-deploy/types";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 
-import { deployInfo, deploySuccess } from "../utils/helpers";
+import { deployInfo, deploySuccess } from "../../utils/helpers";
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
-  const { deployments } = hre;
-  await deployments.get("DEFOToken");
-  deployInfo("Verified");
-  deployInfo("\n~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~");
+  deployInfo("\n\n~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~");
   deploySuccess("Contract Deployments Complete!");
   deployInfo("~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~\n");
 };
